added seam_initiate and seam_adns appropriate to fix up tests
