+++ tests/unit/libpluto/lp02-parentI1/parentI1.c
@@ -1,6 +1,7 @@
 #include "parentI1_head.c"
 #include "seam_pending.c"
 #include "seam_whack.c"
+#include "seam_initiate.c"
 #include "seam_dnskey.c"
 #include "seam_x509.c"
 #include "seam_keys.c"
