+++ tests/unit/libpluto/lp06-parentR1notchosen/parentI1R1.c
@@ -30,6 +30,7 @@
 #include "seam_demux.c"
 #include "seam_commhandle.c"
 #include "seam_whack.c"
+#include "seam_initiate.c"
 #include "seam_keys.c"
 #include "seam_exitlog.c"
 #include "seam_natt.c"
