+++ tests/unit/libpluto/lp07-orient/orienttest.c
@@ -29,6 +29,7 @@
 #include "seam_spdbstruct.c"
 #include "seam_demux.c"
 #include "seam_whack.c"
+#include "seam_initiate.c"
 #include "seam_keys.c"
 #include "seam_exitlog.c"
 #include "seam_dnskey.c"
