+++ tests/unit/libpluto/lp08-parentR1/parentR1_head.c
@@ -28,6 +28,7 @@
 #include "seam_demux.c"
 #include "seam_commhandle.c"
 #include "seam_whack.c"
+#include "seam_initiate.c"
 #include "seam_keys.c"
 #include "seam_exitlog.c"
 #include "seam_natt.c"
