+++ tests/unit/libpluto/lp10-parentI2/parentI2_head.c
@@ -29,6 +29,7 @@
 #include "seam_demux.c"
 #include "seam_commhandle.c"
 #include "seam_whack.c"
+#include "seam_initiate.c"
 #include "seam_exitlog.c"
 #include "seam_natt.c"
 #include "seam_dnskey.c"
