+++ tests/unit/libpluto/lp12-parentR2/parentR2_head.c
@@ -31,6 +31,7 @@
 #include "seam_demux.c"
 #include "seam_commhandle.c"
 #include "seam_whack.c"
+#include "seam_initiate.c"
 #include "seam_exitlog.c"
 #include "seam_natt.c"
 #include "seam_dnskey.c"
