+++ tests/unit/libpluto/lp13-parentI3/parentI3.c
@@ -30,6 +30,7 @@
 #include "seam_demux.c"
 #include "seam_commhandle.c"
 #include "seam_whack.c"
+#include "seam_initiate.c"
 #include "seam_exitlog.c"
 #include "seam_natt.c"
 #include "seam_dnskey.c"
