+++ tests/unit/libpluto/lp27-dnsloadI1/dnscpeI1.c
@@ -2,6 +2,7 @@
 #include "seam_x509.c"
 #include "seam_pending.c"
 #include "seam_whack.c"
+#include "seam_initiate.c"
 #include "seam_keys.c"
 #include "seam_dnskey.c"
