+++ tests/unit/libpluto/lp29-adns/adnstest.c
@@ -32,6 +32,7 @@
 #include "seam_whack.c"
 #include "seam_exitlog.c"
 #include "seam_hostpair.c"
+#include "seam_adns.c"
 
 const char *progname=NULL;
 int verbose=0;
