+++ tests/unit/libpluto/lp31-IDhostpair/IDhostpair.c
@@ -30,6 +30,7 @@
 #include "seam_delete.c"
 #include "seam_ipsecdoi.c"
 #include "seam_whack.c"
+#include "seam_initiate.c"
 #include "seam_exitlog.c"
 #include "seam_natt.c"
 #include "seam_dnskey.c"
