+++ tests/unit/libpluto/seam_adns.c
@@ -1,0 +1,1 @@
+bool kick_adns_connection(struct connection *c) { return TRUE; }
